Fixed some warnings in MSVC 7.1 (64bits) related to size_t int conversions.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23219 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/instmesh/persist/standard/instmeshldr.cpp
@@ -387,14 +387,14 @@ bool csInstFactorySaver::WriteDown (iBase* obj, iDocumentNode* parent,
       paramsNode->CreateNodeBefore(CS_NODE_ELEMENT, 0);
     numvtNode->SetValue("numvt");
     numvtNode->CreateNodeBefore(CS_NODE_TEXT, 0)
-      ->SetValueAsInt(gfact->GetVertexCount());
+      ->SetValueAsInt((int)gfact->GetVertexCount());
 
     //Write NumTri Tag
     csRef<iDocumentNode> numtriNode = 
       paramsNode->CreateNodeBefore(CS_NODE_ELEMENT, 0);
     numtriNode->SetValue("numtri");
     numtriNode->CreateNodeBefore(CS_NODE_TEXT, 0)
-      ->SetValueAsInt(gfact->GetTriangleCount());
+      ->SetValueAsInt((int)gfact->GetTriangleCount());
 
     size_t i;
     //Write Vertex Tags
